Moved the new button bar from InputPanel to PreviewPanel
